+++ pyogrio/_io.pyx
@@ -50,11 +50,11 @@ log = logging.getLogger(__name__)
 # (index in array is the integer field type)
 FIELD_TYPES = [
     "int32",           # OFTInteger, Simple 32bit integer
-    None,              # OFTIntegerList, List of 32bit integers, not supported
+    "list(int32)",     # OFTIntegerList, List of 32bit integers
     "float64",         # OFTReal, Double Precision floating point
-    None,              # OFTRealList, List of doubles, not supported
+    "list(float64)",   # OFTRealList, List of doubles
     "object",          # OFTString, String of UTF-8 chars
-    None,              # OFTStringList, Array of strings, not supported
+    "list(str)",       # OFTStringList, Array of strings
     None,              # OFTWideString, deprecated, not supported
     None,              # OFTWideStringList, deprecated, not supported
     "object",          # OFTBinary, Raw Binary data
@@ -62,7 +62,7 @@ FIELD_TYPES = [
     None,              # OFTTime, Time, NOTE: not directly supported in numpy
     "datetime64[ms]",  # OFTDateTime, Date and Time
     "int64",           # OFTInteger64, Single 64bit integer
-    None               # OFTInteger64List, List of 64bit integers, not supported
+    "list(int64)"      # OFTInteger64List, List of 64bit integers, not supported
 ]
 
 # Mapping of OGR integer field types to OGR type names
@@ -915,6 +915,10 @@ cdef process_fields(
     cdef int success
     cdef int field_index
     cdef int ret_length
+    cdef int *ints_c
+    cdef int64_t *int64s_c
+    cdef double *doubles_c
+    cdef char **strings_c
     cdef GByte *bin_value
     cdef int year = 0
     cdef int month = 0
@@ -1005,6 +1009,55 @@ cdef process_fields(
                         year, month, day, hour, minute, second, microsecond
                     ).isoformat()
 
+        elif field_type == OFTIntegerList:
+            # According to GDAL doc, this can return NULL for an empty list, which is a
+            # valid result. So don't use check_pointer as it would throw an exception.
+            ints_c = OGR_F_GetFieldAsIntegerList(ogr_feature, field_index, &ret_length)
+            int_arr = np.ndarray(shape=(ret_length,), dtype=np.int32)
+            for j in range(ret_length):
+                int_arr[j] = ints_c[j]
+            data[i] = int_arr
+
+        elif field_type == OFTInteger64List:
+            # According to GDAL doc, this can return NULL for an empty list, which is a
+            # valid result. So don't use check_pointer as it would throw an exception.
+            int64s_c = OGR_F_GetFieldAsInteger64List(
+                ogr_feature, field_index, &ret_length
+            )
+
+            int_arr = np.ndarray(shape=(ret_length,), dtype=np.int64)
+            for j in range(ret_length):
+                int_arr[j] = int64s_c[j]
+            data[i] = int_arr
+
+        elif field_type == OFTRealList:
+            # According to GDAL doc, this can return NULL for an empty list, which is a
+            # valid result. So don't use check_pointer as it would throw an exception.
+            doubles_c = OGR_F_GetFieldAsDoubleList(
+                ogr_feature, field_index, &ret_length
+            )
+
+            double_arr = np.ndarray(shape=(ret_length,), dtype=np.float64)
+            for j in range(ret_length):
+                double_arr[j] = doubles_c[j]
+            data[i] = double_arr
+
+        elif field_type == OFTStringList:
+            # According to GDAL doc, this can return NULL for an empty list, which is a
+            # valid result. So don't use check_pointer as it would throw an exception.
+            strings_c = OGR_F_GetFieldAsStringList(ogr_feature, field_index)
+
+            string_list_index = 0
+            vals = []
+            if strings_c != NULL:
+                # According to GDAL doc, the list is terminated by a NULL pointer.
+                while strings_c[string_list_index] != NULL:
+                    val = strings_c[string_list_index]
+                    vals.append(get_string(val, encoding=encoding))
+                    string_list_index += 1
+
+            data[i] = np.array(vals)
+
 
 @cython.boundscheck(False)  # Deactivate bounds checking
 @cython.wraparound(False)   # Deactivate negative indexing.
@@ -1048,16 +1101,16 @@ cdef get_features(
     field_indexes = fields[:, 0]
     field_ogr_types = fields[:, 1]
 
-    field_data = [
-        np.empty(
-            shape=(num_features, ),
-            dtype = (
-                "object"
-                if datetime_as_string and fields[field_index, 3].startswith("datetime")
-                else fields[field_index, 3]
-            )
-        ) for field_index in range(n_fields)
-    ]
+    field_data = []
+    for field_index in range(n_fields):
+        if datetime_as_string and fields[field_index, 3].startswith("datetime"):
+            dtype = "object"
+        elif fields[field_index, 3].startswith("list"):
+            dtype = "object"
+        else:
+            dtype = fields[field_index, 3]
+
+        field_data.append(np.empty(shape=(num_features, ), dtype=dtype))
 
     field_data_view = [field_data[field_index][:] for field_index in range(n_fields)]